+++ core/orm/index.js
@@ -14,7 +14,7 @@ const { createSeeder } = require('./seeder');
 const { createScopeContext } = require('./scopes');
 const { ModelEvents, Hooks, HookCancellationError, createEventContext } = require('./events');
 const { omitHiddenColumns, sanitizeForOutput } = require('./utils');
-const { generateNanoid } = require('./utils/nanoid');
+const { generateNanoid, zodNanoid, extendZ } = require('./utils/nanoid');
 
 /**
  * Create a database instance
@@ -275,6 +275,8 @@ module.exports = {
   extractColumnsFromSchema,
   getColumnMeta,
   generateNanoid,
+  zodNanoid,
+  extendZ,
   // Output sanitization (exclude hidden columns from API/templates)
   omitHiddenColumns,
   sanitizeForOutput,